framework/requirements.brain.txt
asyncio-nats-client
aioredis
jsonschema
pymongo
pyyaml
